requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.2.0
billiard==4.1.0
black==23.11.0
blinker==1.7.0
boto3==1.28.53
botocore==1.31.53
Brotli==1.1.0
build==1.2.1
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.7
dnslib==0.9.25
dnspython==2.6.1
dramatiq==1.15.0
dramatiq-sqs==0.2.0
ecdsa==0.19.0
exceptiongroup==1.1.3
fabric==3.2.2
fake_user_agent==2.1.8
filelock==3.12.4
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
invoke==2.2.0
jaconv==0.3.4
jmespath==1.0.1
kaitaistruct==0.10
kombu==5.3.2
localstack==3.7.1
localstack-core==3.7.1
localstack-ext==3.7.1
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
numpy==2.1.2
outcome==1.2.0
packaging==23.1
paramiko==3.4.0
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
plux==1.11.1
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==6.0.0
pyaes==1.6.1
pyasn1==0.5.1
pycparser==2.21
Pygments==2.18.0
pykakasi==2.2.1
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pyparsing==3.1.1
pyproject-api==1.6.1
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.2
redis==4.6.0
requests==2.31.0
rich==13.8.0
rsa==4.9
s3transfer==0.6.2
selenium==4.8.3
selenium-wire==5.1.0
semver==3.0.2
sentry-sdk==2.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tabulate==0.9.0
tailer==0.4.1
tomli==2.0.1
tox==4.11.3
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.6
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
zstandard==0.22.0
